[Scanner] Learns to parse nested-multiline C-style comments.
The scanner now ignores everything enclosed within `/*` and `*/`,
known as C-style comments, including new line characters. As well,
in case of nested comments, the scanner makes sure that the comments
are properly closed, meaning that the following code:

```lox
/* this is a comment /* but we started a new one
and we close the first one here */
```

will trigger an error, as there is a mismatched comment, i.e. a `*/`
is missing.