[AST] First version of Expressions meta-programming.
With the help of sourcery we define the first definition of a template
that will generate classes that define the subclasses of expressions
that we support.

`Expr` is the "abstract class" and `Binary`, `Grouping`,
`Literal` and `Unary` derive from it.

They are supposed to be static classes that will simply host methods,
so maybe they can be instead enum with associated values plus
method definitions.