[Scanner] `TokenType` no associated value and `Literal` with optional value.
This introduces two changes in how tokens are handled.

1. `TokenType` has no more associated values: Literals, Strings and Number
tokens had associated values but that does not allow for comparing
two Tokens based on the token type without taking into account the
associated value, i.e. it was not possible to compare if two given tokens
were Strings without looking at the associated literal string.
This is specially key when writting parsing code.

2. In Lox, we can use `nil`. During scanning, it is represented by
a `Literal` token. As `Literal` can represent a String and a Number,
the value has type `AnyObject`. Unfortunately in swift, `nil` can't be
casted to `AnyObject`, hence the value attached to a literal is from
now on, `Optional(AnyObject)`.