Refactoring error handling in the stack.
When executing the interpretation phase of the compiler,
we needed to propagate the errors we throw up the stack
to be able to handle them properly.

In order to surface the errors we threw we needed to mark
all functions that evaluate an expression as *throwing*
(suffix `throw` added to functions) and since we are using
the Visitor pattern based on protocols, a new protocol was defined
that marks all possible visit functions marked as throwing.

As well, we centralized everything related to errors (enums
and utilities) in `ErrorUtil.swift` and there is an enum
that encapsulates all possible errors in all phases of the compiler.

This new way of handling exceptions when visiting and evaluating
the AST simply how we handle errors when executing files
and in the REPL.