[parse] Avoiding name clash for associated types in Visitor protocols.
We recently introduced visitor protocol for statements but the
associated type that will be returned by each visit function
was left as `Return` inducing a name clash when we want to make
the `Interpreter` class to be able to navigate `Expr`s and `Stmt`s
at the same time (by conforming to both visitor protocols).

Therefore we only way to avoid the clash is to rename the associated
type.