[eval] Introducing a container to keep track of variable values.
`Environment` encapsulates a dictionary to keep track of
values assigned to variables. For the time being the dictionary
is populated whenever a variable is defined. If the variable
definition includes an expression then it is evaluated and the
result is put in as value. Otherwise the entry contains `nil`.

`var a = 32;` will put in the dictionary the entry `a -> 32`.
`var a;` will put in the dictionary the entry `nil`.

Variable re-definition is allowed for now and we can not change
the value of a variable.